Create CompOffRemainder.gs
For those of us in service, there may be compensatory offs. In my place the these compensatory offs (leaves) will have to be awailed within a 6 month period, else will get expired. 

So enter the year, month, day when you actually worked, day of expiry is calculated. As and when the leave is awailed that row from the spreadsheet may be deleted.

A email is sent to a desired email id every 15 days which remainds of the current status of the leaves.